backend/requirements.txt
flask
python-dotenv
Flask-RESTful